Move span operations into span_ops header
